管理端前端部分第二次迭代完成

+++ manager_frontend/src/views/Notice.vue
@@ -5,15 +5,13 @@
     </el-dialog>
     <div class="container">
       <div class="toolbar">
-        <Toolbar :show-nums="showNums" :query-start-date="queryStartDate" :query-end-date="queryEndDate"
-                 :query-str="queryStr" choose-num choose-date query
-                 @showNumsChange="showNumsChange" @queryStartDateChange="queryStartDateChange"
-                 @queryEndDateChange="queryEndDateChange" @queryStrChange="queryStrChange" @query="query">
+        <Toolbar :toolbar="toolbar" choose-num choose-date query
+                 @query="query">
         </Toolbar>
         <el-button type="primary" plain @click="handleCreate">创建公告</el-button>
       </div>
       <el-table class="table" :data="slicedData" v-loading="tableLoading" @sort-change="doSort"
-                :default-sort="{prop: 'postDate', order: 'descending'}">
+                :default-sort="{prop: 'noticeID', order: 'descending'}">
         <el-table-column type="index" width="50"></el-table-column>
         <el-table-column prop="postDate" label="发布时间" width="150" :sortable="'custom'"></el-table-column>
         <el-table-column prop="expireDate" label="过期时间" width="150" :sortable="'custom'"></el-table-column>
@@ -39,8 +37,8 @@
 </template>
 
 <script>
-import NoticeEditor from '../components/NoticeEditor'
-import Toolbar from '../components/Toolbar'
+import NoticeEditor from '@/components/NoticeEditor'
+import Toolbar from '@/components/Toolbar'
 
 export default {
   name: "Notice",
@@ -62,12 +60,14 @@ export default {
         prop: 'noticeID',
         order: 'descending'
       },
+      toolbar: {
+        showNums: 20,
+        queryStartDate: null,
+        queryEndDate: null,
+        queryStr: '',
+      },
       pageSize: 10,
       currentPage: 1,
-      showNums: 20,
-      queryStartDate: null,
-      queryEndDate: null,
-      queryStr: '',
       editNoticeID: null,
       showNoticeEditor: false,
       form: {
@@ -99,54 +99,42 @@ export default {
         this.sortType.prop = 'noticeID'
         this.sortType.order = 'descending'
       }
-      this.noticeList.sort((a, b) => {
-        const var1 = a[this.sortType.prop]
-        const var2 = b[this.sortType.prop]
-        const asc = this.sortType.order === 'ascending'
-        if (var1 < var2) {
-          return asc ? -1 : 1
-        } else if (var1 > var2) {
-          return asc ? 1 : -1
-        } else {
-          if (a['noticeID'] < b['noticeID']) {
-            return asc ? -1 : 1
-          } else if (a['noticeID'] > b['noticeID']) {
-            return asc ? 1 : -1
-          } else {
-            return 0
-          }
-        }
-      })
+      this.$utils.sort(this.noticeList, this.sortType, 'noticeID')
     },
     getNoticeList() {
       let params = {
-        num: this.showNums,
-        queryStartDate: this.queryStartDate && this.$utils.time.format(this.queryStartDate, 'yyyy-MM-dd'),
-        queryEndDate: this.queryEndDate && this.$utils.time.format(this.queryEndDate, 'yyyy-MM-dd'),
-        queryStr: this.queryStr
+        num: this.toolbar.showNums
       }
-      if (this.queryStartDate && this.queryEndDate && this.queryStr) {
+      if (this.toolbar.queryStartDate && this.toolbar.queryEndDate && this.toolbar.queryStr) {
+        if (this.toolbar.queryStartDate > this.toolbar.queryEndDate) {
+          this.$utils.alertMessage(this, '请选择正确的时间区间', 'warning')
+          return
+        }
         params = {
           ...params,
+          queryStartDate: this.toolbar.queryStartDate && this.$utils.time.format(this.toolbar.queryStartDate, 'yyyy-MM-dd'),
+          queryEndDate: this.toolbar.queryEndDate && this.$utils.time.format(this.toolbar.queryEndDate, 'yyyy-MM-dd'),
+          queryStr: this.toolbar.queryStr,
           queryType: 3
         }
-      } else if (this.queryStartDate && this.queryEndDate) {
-        delete params.queryStr
+      } else if (this.toolbar.queryStartDate && this.toolbar.queryEndDate) {
+        if (this.toolbar.queryStartDate > this.toolbar.queryEndDate) {
+          this.$utils.alertMessage(this, '请选择正确的时间区间', 'warning')
+          return
+        }
         params = {
           ...params,
+          queryStartDate: this.toolbar.queryStartDate && this.$utils.time.format(this.toolbar.queryStartDate, 'yyyy-MM-dd'),
+          queryEndDate: this.toolbar.queryEndDate && this.$utils.time.format(this.toolbar.queryEndDate, 'yyyy-MM-dd'),
           queryType: 1
         }
-      } else if (this.queryStr) {
-        delete params.queryStartDate
-        delete params.queryEndDate
+      } else if (this.toolbar.queryStr) {
         params = {
           ...params,
+          queryStr: this.toolbar.queryStr,
           queryType: 2
         }
       } else {
-        delete params.queryStartDate
-        delete params.queryEndDate
-        delete params.queryStr
         params = {
           ...params,
           queryType: 0
@@ -157,6 +145,7 @@ export default {
         if (res.data.errCode === 0) {
           this.noticeList = res.data.noticeList
           this.doSort()
+          this.$utils.alertMessage(this, '获取数据成功', 'success')
         } else {
           this.$utils.error.APIError(this, res.data)
         }
@@ -165,21 +154,6 @@ export default {
         this.$utils.error.ServerError(this, err)
         this.tableLoading = false
       })
-    },
-    showNumsChange(val) {
-      this.showNums = val
-      this.getNoticeList()
-    },
-    queryStartDateChange(val) {
-      this.queryStartDate = val
-      this.getNoticeList()
-    },
-    queryEndDateChange(val) {
-      this.queryEndDate = val
-      this.getNoticeList()
-    },
-    queryStrChange(val) {
-      this.queryStr = val
     },
     query() {
       this.getNoticeList()
@@ -203,18 +177,25 @@ export default {
       this.showNoticeEditor = true
     },
     handleDelete(row) {
-      this.tableLoading = true
-      this.$api.notice.deleteNotice({noticeID: row.noticeID}).then((res) => {
-        if (res.data.errCode === 0) {
-          this.$utils.alertMessage(this, '删除成功', 'success')
-          this.getNoticeList()
-        } else {
-          this.$utils.error.APIError(this, res.data)
+      this.$confirm('此操作将永久删除该公告, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.tableLoading = true
+        this.$api.notice.deleteNotice({noticeID: row.noticeID}).then((res) => {
+          if (res.data.errCode === 0) {
+            this.$utils.alertMessage(this, '删除成功', 'success')
+            this.getNoticeList()
+          } else {
+            this.$utils.error.APIError(this, res.data)
+            this.tableLoading = false
+          }
+        }).catch((err) => {
+          this.$utils.error.ServerError(this, err)
           this.tableLoading = false
-        }
-      }).catch((err) => {
-        this.$utils.error.ServerError(this, err)
-        this.tableLoading = false
+        })
+      }).catch(() => {
       })
     },
     editorCancel() {
@@ -234,6 +215,10 @@ export default {
         this.$api.notice.createNotice(params).then((res) => {
           if (res.data.errCode === 0) {
             this.$utils.alertMessage(this, '创建成功', 'success')
+            this.form = {
+              expireDate: null,
+              noticeContent: '',
+            }
             this.getNoticeList()
           } else {
             this.$utils.error.APIError(this, res.data)
